requirements.txt
# Core scientific computing
numpy==1.24.3
scipy==1.11.1
pandas==2.0.3

# Image processing
Pillow==10.0.0
scikit-image==0.21.0
opencv-python==4.8.0.76

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Deep learning
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0

# Whole slide imaging
openslide-python==1.3.1

# Segmentation
cellpose==3.0.1
stardist==0.8.5
csbdeep==0.7.4

# Machine learning
scikit-learn==1.3.0
umap-learn==0.5.3

# Image registration (for multi-modal analysis)
SimpleITK>=2.2.0

# Utilities
tqdm==4.65.0
